+++ x/msgfees/types/codec.go
@@ -1,6 +1,8 @@
 package types
 
 import (
+	proto "github.com/gogo/protobuf/proto"
+
 	"github.com/cosmos/cosmos-sdk/codec"
 	cdctypes "github.com/cosmos/cosmos-sdk/codec/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -9,8 +11,13 @@ import (
 
 // ignoring RegisterLegacyAminoCodec registers all the necessary types and interfaces for the
 // double check
-
 func RegisterInterfaces(registry cdctypes.InterfaceRegistry) {
+	messages := make([]proto.Message, len(allRequestMsgs))
+	for i, msg := range allRequestMsgs {
+		messages[i] = msg
+	}
+	registry.RegisterImplementations((*sdk.Msg)(nil), messages...)
+
 	registry.RegisterImplementations(
 		(*govtypesv1beta1.Content)(nil),
 		&AddMsgFeeProposal{},
@@ -18,10 +25,6 @@ func RegisterInterfaces(registry cdctypes.InterfaceRegistry) {
 		&RemoveMsgFeeProposal{},
 		&UpdateNhashPerUsdMilProposal{},
 		&UpdateConversionFeeDenomProposal{},
-	)
-	registry.RegisterImplementations(
-		(*sdk.Msg)(nil),
-		&MsgAssessCustomMsgFeeRequest{},
 	)
 }